Refactor: RefreshToken의 필드 변경으로 인한 리팩토링
user_email -> email
refresh_token -> token
